Adding a bootstrap_ec2 helper/script
This follows the same pattern as the clusto commands but, since this
script is meant to be run once (how many times you want to bootstrap
your amazon objects, really?) is not part of the scripts/commands
list. You can, however, run this script many times and it will continue
to work, so if a new region/zone is added to amazon you can use it to
refresh your data.