Revamped the security group stuff and vpc bits
Turns out it didn't quite work the way it was. And security group names
are not unique between ec2 and vpc. So have to make the distinction
between the two of them and still send one of the two depending if it is
ec2 or vpc. Quite a mess really.